Change float to double

+++ backend/src/main/java/trainingjournal/backend/model/Exercise.java
@@ -8,7 +8,7 @@ public record Exercise(
         String description,
         int repeats,
         int sets,
-        float weight
+        double weight
 ) {
 
 }
